+++ packages/utils/src/getWidget.tsx
@@ -1,9 +1,8 @@
 import { createElement } from 'react';
-import get from 'lodash/get';
-import set from 'lodash/set';
 import ReactIs from 'react-is';
 
 import getSchemaType from './getSchemaType';
+import { getByPath, setByPath } from './pathUtils';
 import type { FormContextType, RJSFSchema, Widget, RegistryWidgetsType, StrictRJSFSchema } from './types';
 
 /** The map of schema types to widget type to widget name
@@ -71,13 +70,13 @@ const widgetMap: Record<string, Record<string, string>> = {
 function mergeWidgetOptions<T = any, S extends StrictRJSFSchema = RJSFSchema, F extends FormContextType = any>(
   AWidget: Widget<T, S, F>,
 ) {
-  let MergedWidget: Widget<T, S, F> | undefined = get(AWidget, 'MergedWidget');
+  let MergedWidget: Widget<T, S, F> | undefined = getByPath(AWidget, 'MergedWidget');
   // cache return value as property of widget for proper react reconciliation
   if (!MergedWidget) {
     // oxlint-disable-next-line typescript/no-deprecated
     const defaultOptions = AWidget.defaultProps?.options || {};
     MergedWidget = ({ options, ...props }) => <AWidget options={{ ...defaultOptions, ...options }} {...props} />;
-    set(AWidget, 'MergedWidget', MergedWidget);
+    setByPath(AWidget, 'MergedWidget', MergedWidget);
   }
   return MergedWidget;
 }