+++ packages/utils/src/lookupFromFormContext.ts
@@ -1,7 +1,5 @@
-import get from 'lodash/get';
-import has from 'lodash/has';
-
 import { FORM_CONTEXT_NAME, LOOKUP_MAP_NAME } from './constants';
+import { getByPath, hasByPath } from './pathUtils';
 import type { FormContextType, RJSFSchema, Registry, StrictRJSFSchema } from './types';
 
 /** Given a React JSON Schema Form registry or formContext object, return the value associated with `toLookup`. This
@@ -17,10 +15,10 @@ export default function lookupFromFormContext<
   T = any,
   S extends StrictRJSFSchema = RJSFSchema,
   F extends FormContextType = any,
->(regOrFc: Registry<T, S, F> | Registry<T, S, F>['formContext'], toLookup: string, fallback?: unknown) {
+>(regOrFc: Registry<T, S, F> | Registry<T, S, F>['formContext'], toLookup: string, fallback?: any) {
   const lookupPath = [LOOKUP_MAP_NAME];
-  if (has(regOrFc, FORM_CONTEXT_NAME)) {
+  if (hasByPath(regOrFc, FORM_CONTEXT_NAME)) {
     lookupPath.unshift(FORM_CONTEXT_NAME);
   }
-  return get(regOrFc, [...lookupPath, toLookup], fallback);
+  return getByPath(regOrFc, [...lookupPath, toLookup], fallback);
 }