+++ packages/utils/src/schema/findFieldInSchema.ts
@@ -1,7 +1,5 @@
-import get from 'lodash/get';
-import has from 'lodash/has';
-
 import { ANY_OF_KEY, ONE_OF_KEY, PROPERTIES_KEY, REQUIRED_KEY } from '../constants';
+import { getByPath, hasByPath } from '../pathUtils';
 import type {
   Experimental_CustomMergeAllOf,
   FormContextType,
@@ -60,7 +58,7 @@ export default function findFieldInSchema<
         {} as S,
         experimental_customMergeAllOf,
       );
-      if (has(parentField, ONE_OF_KEY)) {
+      if (hasByPath(parentField, ONE_OF_KEY)) {
         // if this sub-path has a `oneOf` then use the formData to drill into the schema with the selected option
         parentField = findSelectedOptionInXxxOf(
           validator,
@@ -68,10 +66,10 @@ export default function findFieldInSchema<
           parentField,
           fieldNameKey,
           ONE_OF_KEY,
-          get(formData, subPath),
+          getByPath(formData, [subPath]),
           experimental_customMergeAllOf,
         )!;
-      } else if (has(parentField, ANY_OF_KEY)) {
+      } else if (hasByPath(parentField, ANY_OF_KEY)) {
         // if this sub-path has a `anyOf` then use the formData to drill into the schema with the selected option
         parentField = findSelectedOptionInXxxOf(
           validator,
@@ -79,14 +77,14 @@ export default function findFieldInSchema<
           parentField,
           fieldNameKey,
           ANY_OF_KEY,
-          get(formData, subPath),
+          getByPath(formData, [subPath]),
           experimental_customMergeAllOf,
         )!;
       }
     });
   }
 
-  if (has(parentField, ONE_OF_KEY)) {
+  if (hasByPath(parentField, ONE_OF_KEY)) {
     // When oneOf is in the root schema, use the formData to drill into the schema with the selected option
     parentField = findSelectedOptionInXxxOf(
       validator,
@@ -97,7 +95,7 @@ export default function findFieldInSchema<
       formData,
       experimental_customMergeAllOf,
     )!;
-  } else if (has(parentField, ANY_OF_KEY)) {
+  } else if (hasByPath(parentField, ANY_OF_KEY)) {
     // When anyOf is in the root schema, use the formData to drill into the schema with the selected option
     parentField = findSelectedOptionInXxxOf(
       validator,